# perf(weave): late aggregation after limiting in calls query