## Description

<!--
Include reference to internal ticket "Fixes WB-NNNNN" and/or GitHub issue "Fixes #NNNN" (if applicable)
-->

Remove the filtered_calls CTE in favor of direct subquery with an additional subquery, to force resolution of aggregation functions **after** taking the limit. This should not affect the results of queries, but dramatically improves memory util.

Inspired by [this github thread](https://github.com/ClickHouse/ClickHouse/issues/5329)

### Calls query
Prod:
```sql

```

Branch:
```sql

```

### Call stats query
Prod:
```sql
SELECT count() FROM (WITH filtered_calls AS
  (SELECT calls_merged.id AS id
   FROM calls_merged
   WHERE calls_merged.project_id = 'UHJvamVjdEludGVybmFsSWQ6NDEwNDM2MzM='
     AND (calls_merged.parent_id IS NULL)
   GROUP BY (calls_merged.project_id,
             calls_merged.id)
   HAVING (((any(calls_merged.deleted_at) IS NULL))
           AND ((NOT ((any(calls_merged.started_at) IS NULL))))))
SELECT calls_merged.id AS id
FROM calls_merged
WHERE calls_merged.project_id = '<>'
  AND (calls_merged.id IN filtered_calls)
  AND ((calls_merged.inputs_dump LIKE '%"%value%"%'
        OR calls_merged.inputs_dump IS NULL))
GROUP BY (calls_merged.project_id, calls_merged.id)
HAVING position(JSON_VALUE(any(calls_merged.inputs_dump), '$."state"."key"'), 'value') > 0)
```

Branch:
```sql
SELECT count() FROM (
  SELECT calls_merged.id AS id
  FROM calls_merged
  WHERE (calls_merged.project_id, calls_merged.id) IN (
      SELECT project_id, id
      FROM (
          SELECT calls_merged.project_id, calls_merged.id
          FROM calls_merged
          WHERE
              calls_merged.project_id = '<>'
              AND calls_merged.parent_id IS NULL
              AND (calls_merged.inputs_dump LIKE '%"%value"%' OR calls_merged.inputs_dump IS NULL)
          GROUP BY calls_merged.project_id, calls_merged.id
          HAVING
              any(calls_merged.deleted_at) IS NULL AND
              NOT any(calls_merged.started_at) IS NULL AND
              position(JSON_VALUE(any(calls_merged.inputs_dump), '$."key"'), 'value') > 0
          ORDER BY project_id, id))
  GROUP BY calls_merged.project_id, calls_merged.id)
```


## Testing

Calls query:
|                     | Avg. Read Rows | Avg. Memory Usage | Latency |
|---------------------|----------------|--------------------|---------|
| prod                | 2,082,326      | 32.68 GiB          | 8s      |
| branch              | 1,069,748.33   | 2.44 GiB           | 4s      |

Stats query:
|           | Elapsed Time | Rows Read     | Data Read | Peak Memory Usage |
|-----------|--------------|---------------|-----------|-------------------|
| prod      | 15.802s      | 2,083,612      | 44.75 GB  | 64 GiB            |
| branch    | 4.439s       | 1,100,868      | 44.55 GB  | 2.75 GiB          |